Automatic configuration of Config object with HTMLInputElements
More refactoring of configuration, it can now automatically detect which
input fields are present inside the #configmenu div and will add them.

You can also define some data-* attributes which will configure the
parameter:

* data-config: the internal name which will be available on the Config object
* data-extract: the method used to read/write the value in/out of the input field (use valueAsNumber for numbers)
* data-default: default value for said configuration variable

In the future it might be wise to iterate just over HTMLElements which
have a data-config attribute instead of currently relying on it being an
input field.

Also the checking of such fields is only done during construction of the
Config object, might want to move this to a method and emit some events.

Event emitting is now mostly setup for external referencing

* save => emitted just before the localStorage is updated
* load => emitted just after the configuration is loaded AND has been set on Config (so it will have emitted the update:* events)
* update:<variable> => emitted when the setter of a variable has been called